Module1-RAG_Assistant/requirements.txt
streamlit
langchain_community
langchain-openai
faiss-cpu
openai
python-dotenv
rich
beautifulsoup4
requests 